engine: extract per-variable lowering into db_var_fragment
Move the per-variable lowering half of compile_var_fragment out of
db.rs into a new db_var_fragment.rs as a plain, side-effect-free
shared function (lower_var_fragment) that returns its diagnostics as
data rather than accumulating them via salsa.

This is a behavior-preserving refactor with no functional change. It
creates a clean reuse surface so upcoming read-only structural
analysis (the element-level SCC work for the C-LEARN circular
dependency) can run the engine's own per-variable production lowering
without re-deriving it from a reconstructed context, and it keeps
db.rs under the per-file line cap. The salsa-tracked caller in db.rs
replays the returned diagnostics, so accumulation order and all
observable compile output are unchanged.
